Исправление замечаний от CodeFactor
